requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
backports.cached-property==1.0.1
bcrypt==3.2.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
fastapi==0.75.0
graphql-core==3.2.0
h11==0.13.0
idna==3.3
iso8601==0.1.16
mypy==0.941
mypy-extensions==0.4.3
pycparser==2.21
pydantic==1.9.0
Pygments==2.11.2
pypika-tortoise==0.1.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.1
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
strawberry-graphql==0.103.8
tomli==2.0.1
tortoise-orm==0.18.1
typing_extensions==4.1.1
uvicorn==0.17.6
